Moved WorkflowHost extension into it's own file
Added checks to make sure the public api for listeners is only called on the main thread.
